Support ascending Bunco's spectrum hands
Added Bunco's spectrum hands ("bunc_Spectrum" etc) to hand_table, so they can be ascended when played with 6 or more cards. Also added support for Spectrum Five to display as "Spectrum Six" and so on.

(Bunco's own buggy detection of 6+ spectrums is fixed in PR#17 on jumbocarrot0/Bunco, which hasn't been merged as of writing. With that patch and this one, spectrums can be ascended as expected.)